Remove V3 api calls, ovirt and parallel dependencies.
We support now only the v4 RHV api calls

Fixes:
https://bugzilla.redhat.com/show_bug.cgi?id=1751098
