+++ app/models/manageiq/providers/redhat/infra_manager/provision_workflow.rb
@@ -246,8 +246,8 @@ class ManageIQ::Providers::Redhat::InfraManager::ProvisionWorkflow < MiqProvisio
   def filter_allowed_hosts(all_hosts)
     ems = source_ems
     return all_hosts unless ems
-    ovirt_services = ManageIQ::Providers::Redhat::InfraManager::OvirtServices::Builder.new(ems).build(:use_highest_supported_version => true).new(:ems => ems)
-    ovirt_services.filter_allowed_hosts(self, all_hosts)
+    ManageIQ::Providers::Redhat::InfraManager::OvirtServices::V4.new(:ems => ems)
+                                                                .filter_allowed_hosts(self, all_hosts)
   end
 
   def set_or_default_hardware_field_values(vm)
