+++ app/models/manageiq/providers/redhat/infra_manager/vm.rb
@@ -45,9 +45,7 @@ class ManageIQ::Providers::Redhat::InfraManager::Vm < ManageIQ::Providers::Infra
   }.freeze
 
   def provider_object(connection = nil)
-    ovirt_services_class = ManageIQ::Providers::Redhat::InfraManager::OvirtServices::Builder
-                           .build_from_ems_or_connection(:ems => ext_management_system, :connection => connection)
-    ovirt_services_class.new(:ems => ext_management_system).get_vm_proxy(self, connection)
+    ManageIQ::Providers::Redhat::InfraManager::OvirtServices::V4.new(:ems => ext_management_system).get_vm_proxy(self, connection)
   end
 
   def scan_via_ems?
