+++ spec/models/manageiq/providers/redhat/infra_manager/host_spec.rb
@@ -4,12 +4,8 @@ describe ManageIQ::Providers::Redhat::InfraManager::Host do
     let(:ems) { FactoryBot.create(:ems_redhat_with_authentication) }
     subject { FactoryBot.create(:host_redhat, :ems_id => ems.id) }
 
-    before(:each) do
-      allow_any_instance_of(ManageIQ::Providers::Redhat::InfraManager)
-        .to receive(:supported_api_versions).and_return([4])
-    end
-
     it '.supports_quick_stats?' do
+      allow(subject.ext_management_system).to receive(:api_version).and_return('4.2.4')
       expect(subject.supports_quick_stats?).to be true
     end
 
@@ -26,6 +22,7 @@ describe ManageIQ::Providers::Redhat::InfraManager::Host do
     end
 
     it 'calls list on StatisticsService' do
+      allow(subject.ext_management_system).to receive(:api_version).and_return('4.3.6')
       expect_any_instance_of(OvirtSDK4::StatisticsService).to receive(:list)
       subject.quickStats
     end
