+++ spec/models/manageiq/providers/redhat/infra_manager/vm_spec.rb
@@ -91,31 +91,18 @@ describe ManageIQ::Providers::Redhat::InfraManager::Vm do
       end
 
       context "when vm has provider" do
-        let(:ems_redhat) { FactoryBot.create(:ems_redhat) }
-        let(:supported_api_versions) { [3] }
-        let(:vm) { FactoryBot.create(:vm_redhat, :storage => storage) }
+        let(:ems_redhat) { FactoryBot.create(:ems_redhat)}
+        let(:vm) { FactoryBot.create(:vm_redhat, :storage => storage, :ext_management_system => ems_redhat) }
 
-        before(:each) do
-          allow(vm.ext_management_system).to receive(:supported_api_versions).and_return(supported_api_versions)
-
-          context "when provider does not support reconfigure disks" do
-            it "does not support reconfigure disks" do
-              expect(vm.supports_reconfigure_disks?).to be_falsey
-            end
-          end
-
-          context "when provider supports reconfigure disks" do
-            let(:supported_api_versions) { [3] }
-            it "supports reconfigure disks" do
-              expect(vm.supports_reconfigure_disks?).to be_truthy
-            end
-          end
+        it "supports reconfigure disks" do
+          expect(vm.supports_reconfigure_disks?).to be_truthy
         end
       end
     end
   end
 
   describe "#supports_publish?" do
+    let(:ems) { FactoryBot.create(:ems_redhat_with_authentication) }
     context "when vm has no storage" do
       let(:vm) { FactoryBot.create(:vm_redhat, :storage => nil, :ext_management_system => nil) }
 
@@ -135,11 +122,9 @@ describe ManageIQ::Providers::Redhat::InfraManager::Vm do
 
     context "when vm is not in down state" do
       let(:storage) { FactoryBot.create(:storage_nfs, :ems_ref => "http://example.com/storages/XYZ") }
-      let(:ems) { FactoryBot.create(:ems_redhat_with_authentication) }
       let(:vm) { FactoryBot.create(:vm_redhat, :ext_management_system => ems, :storage => storage) }
 
       it "does not support publish" do
-        allow(ems).to receive(:supported_api_versions).and_return([4])
         allow(vm).to receive(:power_state).and_return("on")
 
         expect(vm.supports_publish?).to be_falsey
@@ -148,11 +133,9 @@ describe ManageIQ::Providers::Redhat::InfraManager::Vm do
 
     context "when vm is down" do
       let(:storage) { FactoryBot.create(:storage_nfs, :ems_ref => "http://example.com/storages/XYZ") }
-      let(:ems) { FactoryBot.create(:ems_redhat_with_authentication) }
       let(:vm) { FactoryBot.create(:vm_redhat, :ext_management_system => ems, :storage => storage) }
 
       it "does support publish" do
-        allow(ems).to receive(:supported_api_versions).and_return([4])
         allow(vm).to receive(:power_state).and_return("off")
 
         expect(vm.supports_publish?).to be_truthy
@@ -177,7 +160,6 @@ describe ManageIQ::Providers::Redhat::InfraManager::Vm do
 
     context "v4" do
       it "unregisters a vm via v4 api" do
-        allow(@ems).to receive(:highest_supported_api_version).and_return(4)
         allow(@vm).to receive(:with_provider_object).and_yield(@vm_service)
         allow(@vm_service).to receive(:unregister).and_return(nil)
