+++ spec/models/manageiq/providers/redhat/infra_manager_spec.rb
@@ -63,10 +63,9 @@ describe ManageIQ::Providers::Redhat::InfraManager do
           @vm   = FactoryBot.create(:vm_redhat, :ext_management_system => @ems)
 
           @rhevm_vm_attrs = double('rhevm_vm_attrs')
-          allow(@ems).to receive(:highest_supported_api_version).and_return(4)
           stub_settings_merge(:ems => { :ems_redhat => { :use_ovirt_engine_sdk => use_ovirt_engine_sdk } })
-          @v4_strategy_instance = instance_double(ManageIQ::Providers::Redhat::InfraManager::OvirtServices::Strategies::V4)
-          allow(ManageIQ::Providers::Redhat::InfraManager::OvirtServices::Strategies::V4)
+          @v4_strategy_instance = instance_double(ManageIQ::Providers::Redhat::InfraManager::OvirtServices::V4)
+          allow(ManageIQ::Providers::Redhat::InfraManager::OvirtServices::V4)
             .to receive(:new)
             .and_return(@v4_strategy_instance)
         end
@@ -88,122 +87,6 @@ describe ManageIQ::Providers::Redhat::InfraManager do
         end
       end
     end
-
-    context "version 3" do
-      before do
-        _guid, _server, zone = EvmSpecHelper.create_guid_miq_server_zone
-        @ems  = FactoryBot.create(:ems_redhat_with_authentication, :zone => zone)
-        @hw   = FactoryBot.create(:hardware, :memory_mb => 1024, :cpu_sockets => 2, :cpu_cores_per_socket => 1)
-        @vm   = FactoryBot.create(:vm_redhat, :ext_management_system => @ems)
-
-        @cores_per_socket = 2
-        @num_of_sockets   = 3
-
-        @rhevm_vm_attrs = double('rhevm_vm_attrs')
-        stub_settings_merge(:ems => { :ems_redhat => { :use_ovirt_engine_sdk => false } })
-        allow(@rhevm_vm_attrs).to receive(:fetch_path).with(:name).and_return('myvm')
-        allow(@rhevm_vm_attrs).to receive(:fetch_path).with(:memory).and_return(4.gigabytes)
-        allow(@rhevm_vm_attrs).to receive(:fetch_path).with(:memory_policy, :guaranteed).and_return(2.gigabytes)
-        # TODO: Add tests for when the highest_supported_api_version is 4
-        allow(@ems).to receive(:supported_api_versions).and_return(%w(3))
-        @rhevm_vm = double('rhevm_vm')
-        allow(@rhevm_vm).to receive(:attributes).and_return(@rhevm_vm_attrs)
-        allow(@vm).to receive(:with_provider_object).and_yield(@rhevm_vm)
-      end
-
-      it "cpu_topology=" do
-        spec = {
-          "numCPUs"           => @cores_per_socket * @num_of_sockets,
-          "numCoresPerSocket" => @cores_per_socket
-        }
-
-        expect(@rhevm_vm).to receive(:cpu_topology=).with(:cores => @cores_per_socket, :sockets => @num_of_sockets)
-        @ems.vm_reconfigure(@vm, :spec => spec)
-      end
-
-      it 'updates the current and persistent configuration if the VM is up' do
-        spec = {
-          'memoryMB' => 8.gigabytes / 1.megabyte
-        }
-        allow(@rhevm_vm_attrs).to receive(:fetch_path).with(:status, :state).and_return('up')
-        expect(@rhevm_vm).to receive(:update_memory).with(8.gigabytes, 2.gigabytes, :next_run => true)
-        expect(@rhevm_vm).to receive(:update_memory).with(8.gigabytes, nil)
-        @ems.vm_reconfigure(@vm, :spec => spec)
-      end
-
-      it 'updates only the persistent configuration when the VM is down' do
-        spec = {
-          'memoryMB' => 8.gigabytes / 1.megabyte
-        }
-        allow(@rhevm_vm_attrs).to receive(:fetch_path).with(:status, :state).and_return('down')
-        expect(@rhevm_vm).to receive(:update_memory).with(8.gigabytes, 2.gigabytes)
-        @ems.vm_reconfigure(@vm, :spec => spec)
-      end
-
-      it 'adjusts the increased memory to the next 256 MiB multiple if the VM is up' do
-        spec = {
-          'memoryMB' => 8.gigabytes / 1.megabyte + 1
-        }
-        adjusted = 8.gigabytes + 256.megabytes
-        allow(@rhevm_vm_attrs).to receive(:fetch_path).with(:status, :state).and_return('up')
-        expect(@rhevm_vm).to receive(:update_memory).with(adjusted, 2.gigabytes, :next_run => true)
-        expect(@rhevm_vm).to receive(:update_memory).with(adjusted, nil)
-        @ems.vm_reconfigure(@vm, :spec => spec)
-      end
-
-      it 'adjusts reduced memory to the next 256 MiB multiple if the VM is up' do
-        spec = {
-          'memoryMB' => 8.gigabytes / 1.megabyte - 1
-        }
-        adjusted = 8.gigabytes
-        allow(@rhevm_vm_attrs).to receive(:fetch_path).with(:status, :state).and_return('up')
-        expect(@rhevm_vm).to receive(:update_memory).with(adjusted, 2.gigabytes, :next_run => true)
-        expect(@rhevm_vm).to receive(:update_memory).with(adjusted, nil)
-        @ems.vm_reconfigure(@vm, :spec => spec)
-      end
-
-      it 'adjusts the guaranteed memory if it is larger than the virtual memory if the VM is up' do
-        spec = {
-          'memoryMB' => 1.gigabyte / 1.megabyte
-        }
-        adjusted = 1.gigabyte
-        allow(@rhevm_vm_attrs).to receive(:fetch_path).with(:status, :state).and_return('up')
-        expect(@rhevm_vm).to receive(:update_memory).with(1.gigabyte, adjusted, :next_run => true)
-        expect(@rhevm_vm).to receive(:update_memory).with(1.gigabyte, nil)
-        @ems.vm_reconfigure(@vm, :spec => spec)
-      end
-
-      it 'adjusts the guaranteed memory if it is larger than the virtual memory if the VM is down' do
-        spec = {
-          'memoryMB' => 1.gigabyte / 1.megabyte
-        }
-        adjusted = 1.gigabyte
-        allow(@rhevm_vm_attrs).to receive(:fetch_path).with(:status, :state).and_return('down')
-        expect(@rhevm_vm).to receive(:update_memory).with(1.gigabyte, adjusted)
-        @ems.vm_reconfigure(@vm, :spec => spec)
-      end
-
-      it 'cpu_topology from set_number_of_cpus with number of sockets >= 1' do
-        allow(@vm).to receive(:cpu_cores_per_socket).and_return(2)
-        options = { :user_event => "Console Request Action [vm_set_number_of_cpus], VM [#{@vm.name}]", :value => 2 }
-        expect(@rhevm_vm).to receive(:cpu_topology=).with(:sockets => 1)
-        @ems.vm_set_num_cpus(@vm, options)
-      end
-
-      it 'cpu_topology from set_number_of_cpus with number of sockets < 1' do
-        allow(@vm).to receive(:cpu_cores_per_socket).and_return(2)
-        options = { :user_event => "Console Request Action [vm_set_number_of_cpus], VM [#{@vm.name}]", :value => 1 }
-        expect(@rhevm_vm).to receive(:cpu_topology=).with(:cores => 1, :sockets => 1)
-        @ems.vm_set_num_cpus(@vm, options)
-      end
-
-      it 'adjusts memory from set_memory' do
-        options = { :user_event => "Console Request Action [vm_set_memory], VM [#{@vm.name}]", :value => 2048 }
-        allow(@rhevm_vm_attrs).to receive(:fetch_path).with(:status, :state).and_return('down')
-        expect(@rhevm_vm).to receive(:update_memory).with(2.gigabytes, 2.gigabytes)
-        @ems.vm_set_memory(@vm, options)
-      end
-    end
   end
 
   context ".make_ems_ref" do
@@ -222,134 +105,38 @@ describe ManageIQ::Providers::Redhat::InfraManager do
     end
   end
 
-  context "api versions" do
-    require 'ovirtsdk4'
+  context "#supports_the_api_version?" do
+    let(:ems) { FactoryBot.create(:ems_redhat) }
 
-    let(:ems) { FactoryBot.create(:ems_redhat_with_authentication) }
-    context 'when parsing database api_version' do
-      let(:ems) { FactoryBot.create(:ems_redhat, :api_version => api_version) }
-      subject(:supported_api_versions) { ems.supported_api_versions }
-      context "version 4.2" do
-        let(:api_version) { "4.2.0" }
-        it 'returns versions 3 and 4' do
-          expect(supported_api_versions).to match_array(%w(3 4))
-        end
-      end
-
-      context "version 3.6.1" do
-        let(:api_version) { "3.6.1" }
-        it 'returns versions 3' do
-          expect(supported_api_versions).to match_array(%w(3))
-        end
-      end
-
-      context "version 4.2.0-0.0.master.20170917124606.gita804ef7.el7.centos" do
-        let(:api_version) { "4.2.0-0.0.master.20170917124606.gita804ef7.el7.centos" }
-        it 'returns versions 3 and 4' do
-          expect(supported_api_versions).to match_array(%w(3 4))
-        end
-      end
-
-      context "version 4.2.1.3" do
-        let(:api_version) { "4.2.1.3" }
-        it 'returns versions 3 and 4' do
-          expect(supported_api_versions).to match_array(%w(3 4))
-        end
-      end
-
-      context "version 4.2.1a" do
-        let(:api_version) { "4.2.1a" }
-        it 'returns versions 3 and 4' do
-          expect(supported_api_versions).to match_array(%w(3 4))
-        end
-      end
-
-      context "version 4.3" do
-        let(:api_version) { "4.3" }
-        it 'returns versions 3 and 4' do
-          expect(supported_api_versions).to match_array(%w(3 4))
-        end
-      end
-
-      context "version 5.5" do
-        let(:api_version) { "5.5" }
-        it 'returns empty array' do
-          expect(supported_api_versions).to match_array([])
-        end
-      end
-    end
-
-    context 'when probing provider directly' do
-      subject(:supported_api_versions) { ems.supported_api_versions }
-      context "#supported_api_versions" do
-        it 'calls the OvirtSDK4::Probe.probe' do
-          expect(OvirtSDK4::Probe).to receive(:probe).and_return([])
-          supported_api_versions
-        end
-
-        it 'properly parses ProbeResults' do
-          allow(OvirtSDK4::Probe).to receive(:probe)
-            .and_return([OvirtSDK4::ProbeResult.new(:version => '3'),
-                         OvirtSDK4::ProbeResult.new(:version => '4')])
-          expect(supported_api_versions).to match_array(%w(3 4))
-        end
-      end
-    end
-
-    describe "#supports_the_api_version?" do
-      it "returns the supported api versions" do
-        allow(ems).to receive(:supported_api_versions).and_return([3])
-        expect(ems.supports_the_api_version?(3)).to eq(true)
-        expect(ems.supports_the_api_version?(6)).to eq(false)
-      end
+    it "only major version 4 are supported" do
+      expect(ems.supports_the_api_version?('4.2.5')).to eq(true)
+      expect(ems.supports_the_api_version?(6)).to eq(false)
     end
   end
 
   context "supported features" do
-    let(:ems) { FactoryBot.create(:ems_redhat) }
-    let(:supported_api_versions) { [3, 4] }
+    let(:ems) { FactoryBot.create(:ems_redhat)}
     context "#process_api_features_support" do
       before(:each) do
         allow(SupportsFeatureMixin).to receive(:guard_queryable_feature).and_return(true)
-        allow(described_class).to receive(:api_features)
-          .and_return('3' => %w(feature1 feature3), '4' => %w(feature2 feature3))
         described_class.process_api_features_support
-        allow(ems).to receive(:supported_api_versions).and_return(supported_api_versions)
       end
 
-      context "no versions supported" do
-        let(:supported_api_versions) { [] }
-        it 'supports the right features' do
-          expect(ems.supports_feature1?).to be_falsey
-          expect(ems.supports_feature2?).to be_falsey
-          expect(ems.supports_feature3?).to be_falsey
+      context "all supported regardless the version" do
+        it "all defined features supported" do
+          described_class::SUPPORTED_FEATURES.each do |f|
+            expect(ems.send("supports_#{f}?")).to be_truthy
+          end
         end
       end
 
-      context "version 3 supported" do
-        let(:supported_api_versions) { [3] }
-        it 'supports the right features' do
-          expect(ems.supports_feature1?).to be_truthy
-          expect(ems.supports_feature2?).to be_falsey
-          expect(ems.supports_feature3?).to be_truthy
-        end
-      end
-
-      context "version 4 supported" do
-        let(:supported_api_versions) { [4] }
-        it 'supports the right features' do
-          expect(ems.supports_feature1?).to be_falsey
-          expect(ems.supports_feature2?).to be_truthy
-          expect(ems.supports_feature3?).to be_truthy
-        end
-      end
-
-      context "all versions supported" do
-        let(:supported_api_versions) { [3, 4] }
-        it 'supports the right features' do
-          expect(ems.supports_feature1?).to be_truthy
-          expect(ems.supports_feature2?).to be_truthy
-          expect(ems.supports_feature3?).to be_truthy
+      context "also with blank api_version" do
+        let(:ems) { FactoryBot.create(:ems_redhat, :api_version => api_version) }
+        let(:api_version) { '' }
+        it "all features supported" do
+          described_class::SUPPORTED_FEATURES.each do |f|
+            expect(ems.send("supports_#{f}?")).to be_truthy
+          end
         end
       end
     end
@@ -408,37 +195,11 @@ describe ManageIQ::Providers::Redhat::InfraManager do
       }
     end
     let(:v4_connection) { double(OvirtSDK4::Connection) }
-    let(:v3_connection) { double(Ovirt::Service) }
-
-    before do
-      allow(v3_connection).to receive(:disconnect)
-    end
 
     it "works with version 4" do
       expect(described_class).to receive(:raw_connect_v4).and_return(v4_connection)
       expect(v4_connection).to receive(:test).with(hash_including(:raise_exception => true))
         .and_return(true)
-
-      described_class.raw_connect(options)
-    end
-
-    it "works with version 3" do
-      expect(described_class).to receive(:raw_connect_v4).and_return(v4_connection)
-      expect(v4_connection).to receive(:test).with(hash_including(:raise_exception => true))
-        .and_raise(OvirtSDK4::Error.new('Something failed'))
-      expect(described_class).to receive(:raw_connect_v3).and_return(v3_connection)
-      expect(v3_connection).to receive(:api).and_return(nil)
-
-      described_class.raw_connect(options)
-    end
-
-    it "always closes the V3 connection" do
-      expect(described_class).to receive(:raw_connect_v4).and_return(v4_connection)
-      expect(v4_connection).to receive(:test).with(hash_including(:raise_exception => true))
-        .and_raise(OvirtSDK4::Error.new('Something failed'))
-      expect(described_class).to receive(:raw_connect_v3).and_return(v3_connection)
-      expect(v3_connection).to receive(:api).and_return(nil)
-      expect(v3_connection).to receive(:disconnect)
 
       described_class.raw_connect(options)
     end
